# komga-rust/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Komga Rust API",
    "version": "0.1.0",
    "description": "Rust implementation of Komga media server for comics/manga. REST API for managing libraries, series, books, read progress, readlists, collections, and more."
  },
  "servers": [
    {
      "url": "/",
      "description": "Default server"
    }
  ],
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "tags": ["Tasks"],
        "summary": "List all tasks",
        "operationId": "listTasks",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of tasks",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskPageDto" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Tasks"],
        "summary": "Delete completed/cancelled tasks",
        "operationId": "deleteCompletedTasks",
        "responses": {
          "204": { "description": "Tasks cleared" }
        }
      }
    },
    "/api/v1/tasks/{id}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get task by ID",
        "operationId": "getTask",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Task details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskDto" }
              }
            }
          },
          "404": { "description": "Task not found" }
        }
      },
      "delete": {
        "tags": ["Tasks"],
        "summary": "Delete a specific task",
        "operationId": "deleteTask",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Task deleted" },
          "404": { "description": "Task not found" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TaskDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "taskType": { "type": "string" },
          "status": { "type": "string" },
          "priority": { "type": "integer" },
          "createdDate": { "type": "string", "format": "date-time" },
          "scheduledDate": { "type": "string", "format": "date-time", "nullable": true },
          "executionStartDate": { "type": "string", "format": "date-time", "nullable": true },
          "executionEndDate": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "TaskPageDto": {
        "type": "object",
        "properties": {
          "content": { "type": "array", "items": { "$ref": "#/components/schemas/TaskDto" } },
          "totalElements": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "number": { "type": "integer" },
          "size": { "type": "integer" },
          "empty": { "type": "boolean" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" },
          "numberOfElements": { "type": "integer" }
        }
      },
...
    }
  },
...
}